Added permissions for managing blog contents with admin, editor and author type users

+++ app/Http/Controllers/PageController.php
@@ -12,7 +12,33 @@ class PageController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth')->except(['show', 'index']);
+        // Permission scheme
+        // -------------------------------------------------
+        //             |  User  | Author | Editor | Admin
+        // create      |  N     | Y      | Y      | Y
+        // store       |  N     | Y      | Y      | Y
+        // edit        |  N     | Y      | Y      | Y
+        // save        |  N     | Y      | Y      | Y
+        // destroy     |  N     | N      | N      | Y
+        // index       |  Y     | Y      | Y      | Y
+        // show        |  Y     | Y      | Y      | Y
+        // -------------------------------------------------
+
+        $this->middleware('admin')->only('destroy');
+
+        $this->middleware(function ($request, $next) {
+
+            if (Auth::user() 
+                && in_array (Auth::user()->type, array('Author', 'Editor', 'Admin'))) 
+            {
+                return $next($request);
+            }
+
+            // if not allowed, then redirect
+            flash ('You do not have permission to perform this action')->warning();
+            return redirect()->back();
+
+        })->only(['create', 'store', 'edit', 'save']);
     }
 
 
@@ -34,8 +60,7 @@ class PageController extends Controller
         
         
     	if ($page->publish == '1' 
-                && ($page->category_id == null 
-                || $page->category->slug === $categorySlug))
+                && ($page->category_id == null || $page->category->slug === $categorySlug))
     	   return view ('page.show', compact('page'));
 
         abort(404, 'Page Not Found');
@@ -44,11 +69,15 @@ class PageController extends Controller
 
     protected function create ()
     {
+        if(! $this->hasPermission ('create')) return redirect()->back();
     	return view ('page.create');
     }
 
     protected function store (Request $request)
-    {        
+    {      
+
+        if(! $this->hasPermission ('store')) return redirect()->back();
+
         $markdown = $request->input('body');
         $markup   = $this->_getHTML($markdown);
     	$doc = new Page ([
@@ -64,15 +93,17 @@ class PageController extends Controller
 
     	$page = Auth::user()->pages()->save($doc);
     	
-    	return redirect()
-    		->route('page-edit', $page->id)
-    		->withMessage('Page saved');
+        flash('Page saved successfully')->success();
+    	return redirect()->route('page-edit', $page->id);
     }
 
 
     protected function edit ($id)
     {
         $page = Page::findOrFail($id);
+
+        if(! $this->hasPermission ('edit', $page)) return redirect()->back();
+
         return view ('page.edit', compact('page'));
     }
 
@@ -82,6 +113,8 @@ class PageController extends Controller
         $id = $request->input('id');
         $page = Page::findOrFail($id);
         
+        if(! $this->hasPermission ('save', $page)) return redirect()->back();
+
         $page->title         = $request->input('head');
         $page->intro         = $request->input('summary');
         $page->category_id   = $request->input('cat');
@@ -93,19 +126,56 @@ class PageController extends Controller
 
         $page->save();
         
-        return redirect()
-            ->route('page-edit', $id)
-            ->withMessage('Page saved');
+        flash('Page saved successfully')->success();
+        return redirect()->route('page-edit', $id);
     }
+
 
     protected function destroy ($id)
     {
+        if(! $this->hasPermission ('delete'))  return redirect()->back();
+
         $page = Page::findOrFail($id);
         $page->delete();
 
-        return redirect()
-            ->route('page-index')
-            ->withMessage('Page saved');
+        flash('Page deleted successfully')->success();
+        return redirect()->route('page-index');
+            
+    }
+
+
+
+    private function hasPermission ($action, $resource = null)
+    {
+            $type = Auth::user()->type;
+
+            // admin should be able to do anything
+            if ($type == 'Admin') {
+                return true;
+            }
+            
+            // editor is like admin, but does not have delete rights
+            if($type == 'Editor' && $action != 'delete') {
+                return true;
+            }
+
+            // authors are like editors, but when it comes to 
+            // edit or update, they can do so only in their
+            // own articles.
+            if($type == 'Author') 
+            {
+                // creation is fine, after all that's what authors do
+                if(in_array($action, array('create', 'store')))
+                    return true;
+
+                // but update is only restricted to own contents
+                if (in_array($action, array('edit', 'save')) && !empty($resource) && $resource->author->id === Auth::user()->id) 
+                    return true;
+            }
+
+            // for everything else, deny permission
+            flash ('You do not have permission to ' . $action . ' this page')->warning();
+            return false;
     }
